refactor(core): own mercurial vcs in internal plugin
